Move shutdown -> thread
